backend/requirements.txt
Django==2.1.1
djangorestframework==3.7.3
mysqlclient==1.3.13
Pillow==5.2.0
pytz==2018.5
